pyproject.toml
[project]
name = "vtc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyannote-audio>=3.4.0",
    "segma",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
segma = { git = "https://github.com/arxaqapi/segma", rev = "1ded41b111856cac128a749015791487e48c5167" }
